Rename multirest.ts to measurerest.ts
